vicharak-config: cli: u-boot-menu: Add support for alternative DTBO path
The mainline dtso files compiled into dtbo with a slightly different
manner, where the first fragment is the metadata node.

Previously metadata node was used directly instead of being a fragment
which was not getting parsed by the overlay manager.

Fix this by checking for the metadata property in the first fragment.

Change-Id: I354495845394269f6d2bd0f60b80814d9d2948db
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>