fix: make PARTUUIDs distinct for each installation
This commit updates the postinst script to randomize PARTUUIDs for all
partitions. This is done inorder to enable mutliboot support for u-boot which
is relying on partuuid given in `/boot/extlinux/extlinux.conf`.

Since all installations are created using the same disk image, they initially
share identical PARTUUIDs. To resolve this, the script ensures that a new,
unique PARTUUID is assigned during the first boot for new image installation or
when updating the vicharak-config package on existing devices.

Signed-off-by: Ajit Singh <ajeetsinghchahar2@gmail.com>